+++ ruzstd/fuzz/fuzz_targets/interop.rs
@@ -35,13 +35,13 @@ fn encode_zstd(data: &[u8]) -> Result<Vec<u8>, std::io::Error> {
 fn encode_ruzstd_uncompressed(data: &mut dyn std::io::Read) -> Vec<u8> {
     let mut input = Vec::new();
     data.read_to_end(&mut input).unwrap();
-    compress_to_vec(data, CompressionLevel::Uncompressed)
+    compress_to_vec(input.as_slice(), CompressionLevel::Uncompressed)
 }
 
 fn encode_ruzstd_compressed(data: &mut dyn std::io::Read) -> Vec<u8> {
     let mut input = Vec::new();
     data.read_to_end(&mut input).unwrap();
-    compress_to_vec(data, CompressionLevel::Uncompressed)
+    compress_to_vec(input.as_slice(), CompressionLevel::Fastest)
 }
 
 fn decode_zstd(data: &[u8]) -> Result<Vec<u8>, std::io::Error> {